Updated pose detection with crop/focus routines
Added bounding box focusing on the call to the posemarker
to improve detection of Poses in instances where the
subject is small in the frame. The initial bounding box
is determined during the orientatino test, where
pose quality is assessed on a frame-by-frame basis
with strides and maximum frame number passed in as
cli args. The bounding box is updated on each frame
where a pose is successfully detected, and upated
with focusing used only on those frames where no
Pose is detected.