First past post orientation check
Added a first-past-post system for checking video
orientation - i.e., looked for frames of sufficient
quality to get a Pose on multiple rotations and ran
through the previously-established quality checks,
which include nose-above-shoulders (probably most informative)
as well as a few other heuristics. This results in
a bounding box hint derived from the first frame
that passes the checks for frames where the models
cannot resolve a Pose. Updated the README with hints
on new parameters that are accessible via CLI calls.