Mocopi ND pilot data and analysis scripts
This has a bunch of functionality for aligning and analyzing
the mocopi data compared to esp32 volum+ synced data
captures. A sunnary is in the dedicated mcocopi_readme,
but I won't integrate (or do a proper readthedocs) yet
as the synchronization needs improvement on the hardware
side which is ongoing as of this commit date. The mocopi
stuff also needs a bit of refactoring.